Fix: Define and encapsulate attributes
Encapsulating the pageCount attribute, it is now changed and
returned by public method, set api key as class attribute
making it easier to find.

The pageCount attribute was being returned and changed directly
no object-oriented, to find the correct place to insert the api
key was necessary to roll the code to the end, now it becomes
easy as class attribute.